+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Automatic SSL/TLS detection on non-standard ports
 - URL Screenshotter agent now takes extra steps to ensure that the browser process is killed after use
+- Version flag to output current version (woah!!!)
 
 ### Changed
 - Packages and other dependencies have been updated to latest versions